lerobot_teleoperator_zhonglin/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "lerobot_teleoperator_zhonglin"
version = "0.1.0"
dependencies = ["lerobot", "pyserial", "numpy"]